Added ignore export capabilities
